Fix for duplicate reply-to headers
There is a bug in Railgun where `build_message_object` would add the `h:reply-to` header, and where `transform_for_mailgun` would also add the `h:Reply-To` header.

This commit makes sure that `transform_for_mailgun` does not add the `h:Reply-To` header again, and this commit changes `reply_to=` in `MessageBuilder` to use the capitalized header (`h:Reply-To`).